+++ CHANGELOG.md
@@ -12,6 +12,10 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 - Configure the PCRE engine with lower match-attempts and recursion limits in order
   to prevent regex matching from potentially "hanging" Semgrep
+- Terraform: Parse heredocs respecting newlines and whitespaces, so that it is
+  possible to correctly match these strings with `metavariable-regex` or
+  `metavariable-pattern`. Previously, Semgrep had problems analyzing e.g. embedded
+  YAML content. (#4582)
 
 ## [0.82.0](https://github.com/returntocorp/semgrep/releases/tag/v0.82.0) - 02-08-2022